requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
gunicorn>=21.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
